+++ Source/SwiftLintFramework/Rules/ReturnArrowWhitespaceRule.swift
@@ -15,8 +15,8 @@ public struct ReturnArrowWhitespaceRule: Rule {
     public let identifier = "return_arrow_whitespace"
 
     public func validateFile(file: File) -> [StyleViolation] {
-        // space doesn't include \n so that "func abc()->\n" can pass validation
-        let space = "[ \\f\\r\\t\\v]"
+        // just horizontal spacing so that "func abc()->\n" can pass validation
+        let space = "[ \\f\\r\\t]"
         let spaceRegex = "(\(space){0}|\(space){2,})"
 
         // ex: func abc()-> Int {
@@ -38,13 +38,14 @@ public struct ReturnArrowWhitespaceRule: Rule {
     public let example = RuleExample(
         ruleName: "Returning Whitespace Rule",
         ruleDescription: "This rule checks whether you have 1 space before " +
-        "return arrow and return type",
+        "return arrow and return type. Newlines are also acceptable.",
         nonTriggeringExamples: [
             "func abc() -> Int {}\n",
             "func abc() -> [Int] {}\n",
             "func abc() -> (Int, Int) {}\n",
             "var abc = {(param: Int) -> Void in }\n",
-            "func abc() ->\n"
+            "func abc() ->\n    Int {}\n",
+            "func abc()\n    -> Int {}\n"
         ],
         triggeringExamples: [
             "func abc()->Int {}\n",